+++ changelog.md
@@ -2,6 +2,7 @@
 
 # Unreleased
 
+* 2024-09-10 Containers may define relinquish tests to dictate when items may be taken from them.
 * 2024-09-08 Verb smart tests onFailure arguments are now varargs.
 * 2024-09-07 Ensure we're exporting types.
 * 2024-09-07 Made containerListing able to be dynamic.